Fix getApplicant method naming
Use better method name "getAdditionalApplicant" in the
CreateApplicationRequest and SalesCreateApplicationRequest and
"getAdditionalApplicants" in Application.

This should make the code less confusing, since it's now clear that
there are indeed ADDITIONAL applicants and that the one that returns an
array has the "s" in the end of the name for plural.

This fixes a bug in SalesCreateApplicationRequest, which used to call
`application->getApplicants()` but the method name was previously
`getApplicant`.